+++ client/src/components/ProductCard/ProductCard.js
@@ -1,18 +1,9 @@
 import React from "react";
 import { Link } from "react-router-dom";
+import { handleAddCart } from "../../utils/addcart";
 import "./ProductCard.css";
 
 export default function ProductCard(props) {
-  const handleAddCart = () => {
-    let oldCart = JSON.parse(localStorage.getItem("cart"));
-    if (oldCart === null) {
-      oldCart = [];
-    }
-    oldCart.push(props.product);
-    localStorage.setItem("cart", JSON.stringify(oldCart));
-    console.log(oldCart);
-  };
-
   return (
     <div className="card-container" key={props.key}>
       <Link to={`/products/${props.id}`}>
@@ -24,7 +15,7 @@ export default function ProductCard(props) {
         <h4 className="card-title">{props.title}</h4>
       </div>
       <div className="group">
-        <button className="add-cart" onClick={handleAddCart}>
+        <button className="add-cart" onClick={handleAddCart(props.product)}>
           Add to cart
         </button>
         <h6>{props.price}</h6>
