+++ client/src/screens/ShoppingCart/ShoppingCart.js
@@ -1,0 +1,8 @@
+import { useEffect } from "react";
+
+export default function ShoppingCart() {
+  useEffect(() => {
+    const cart = JSON.parse(localStorage.getItem("cart"));
+  }, []);
+  return <div></div>;
+}
