+++ client/src/utils/addcart.js
@@ -1,0 +1,9 @@
+export const handleAddCart = (product) => {
+  let oldCart = JSON.parse(localStorage.getItem("cart"));
+  if (oldCart === null) {
+    oldCart = [];
+  }
+  oldCart.push(product);
+  localStorage.setItem("cart", JSON.stringify(oldCart));
+  console.log(oldCart);
+};
